Expand adsbygoogle.js surrogate script
While redirecting requests for the adsbygoogle.js script to an
effectively empty script does often work, some websites break or
detect that the ads were not displayed.

Let's improve on that, by starting with Mozilla's surrogate script[1],
and adding some logic to setup the rough element structure that the
real script creates.

1 - https://searchfox.org/mozilla-central/source/browser/extensions/webcompat/shims/google-ads.js