Add warning about updating mapping.json
It turns out that updating the mapping.json file is kind of
risky. While platforms take the surrogates scripts from the latest
release, mapping.json is taken straight from the HEAD of
main. Redirection rules are then automatically added, sometimes even
when the request would have otherwise been allowlisted. This is made
riskier since if a platform does not have a surrogate script for a
rule yet, the corresponding requests will be blocked instead of
redirected. I've added a warning to the README to hopefully help folks
avoid this trap.